feat(系统管理):  根据权限获取菜单树
根据权限获取菜单树

+++ system/src/main/java/cn/cuiot/dmp/system/infrastructure/persistence/dao/OrgTypeMenuDao.java
@@ -20,7 +20,7 @@ public interface OrgTypeMenuDao {
      * @param orgTypeId
      * @return
      */
-    List<String> getMenuIdListByOrgType(@Param("orgTypeId") Integer orgTypeId);
+    List<String> getMenuIdListByOrgType(@Param("orgTypeId") Long orgTypeId);
 
     /**
      * 根据类型删除菜单
